fix(rating): use ratee association path and Double return type
The AVG() query introduced in f572880 referenced 'r.rateeId' but the
Rating entity has a @ManyToOne 'ratee' association — there is no
'rateeId' property — so Hibernate raised UnknownPathException at
context startup. JPQL AVG() also returns Double, not BigDecimal, so
the method signature was unsatisfiable.

Traverse the association (r.ratee.id) and return Double; RatingService
converts via BigDecimal.valueOf before setScale for the persisted
column. Update the unit test mock to return 5.0d.

+++ backend/src/main/java/com/skillseed/rating/service/RatingService.java
@@ -191,10 +191,11 @@ public class RatingService {
             return;
         }
         // Single aggregate query — no row materialisation. Old code loaded
-        // every rating row into memory just to sum a column.
-        BigDecimal newAvg = ratingRepository.averageOverallScoreForRatee(rateeId);
+        // every rating row into memory just to sum a column. JPQL AVG()
+        // returns Double; convert to BigDecimal for the persisted column.
+        Double newAvg = ratingRepository.averageOverallScoreForRatee(rateeId);
         if (newAvg != null) {
-            ratee.setRatingAvg(newAvg.setScale(2, RoundingMode.HALF_UP));
+            ratee.setRatingAvg(BigDecimal.valueOf(newAvg).setScale(2, RoundingMode.HALF_UP));
         }
         ratee.setSessionsCompleted(ratee.getSessionsCompleted() + 1);
         userRepository.save(ratee);
